Move apiTypes to schemas/
