+++ browser_tests/actionbar.spec.ts
@@ -1,7 +1,7 @@
 import type { Response } from '@playwright/test'
 import { expect, mergeTests } from '@playwright/test'
 
-import type { StatusWsMessage } from '../src/types/apiTypes.ts'
+import type { StatusWsMessage } from '../src/schemas/apiTypes.ts'
 import { comfyPageFixture } from './fixtures/ComfyPage'
 import { webSocketFixture } from './fixtures/ws.ts'
