+++ src/components/dialog/content/ExecutionErrorDialogContent.vue
@@ -63,9 +63,9 @@ import { useI18n } from 'vue-i18n'
 import NoResultsPlaceholder from '@/components/common/NoResultsPlaceholder.vue'
 import FindIssueButton from '@/components/dialog/content/error/FindIssueButton.vue'
 import { useCopyToClipboard } from '@/composables/useCopyToClipboard'
+import type { ExecutionErrorWsMessage, SystemStats } from '@/schemas/apiTypes'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
-import type { ExecutionErrorWsMessage, SystemStats } from '@/types/apiTypes'
 import type { ReportField } from '@/types/issueReportTypes'
 
 import ReportIssuePanel from './error/ReportIssuePanel.vue'
