+++ src/components/sidebar/tabs/nodeLibrary/NodeTreeFolder.vue
@@ -11,9 +11,9 @@
 import { Ref, computed, inject, onMounted, onUnmounted, ref, watch } from 'vue'
 
 import TreeExplorerTreeNode from '@/components/common/TreeExplorerTreeNode.vue'
+import type { BookmarkCustomization } from '@/schemas/apiTypes'
 import { useNodeBookmarkStore } from '@/stores/nodeBookmarkStore'
 import { ComfyNodeDefImpl } from '@/stores/nodeDefStore'
-import type { BookmarkCustomization } from '@/types/apiTypes'
 import { RenderedTreeExplorerNode } from '@/types/treeExplorerTypes'
 
 const props = defineProps<{
