+++ src/composables/widgets/useFloatWidget.ts
@@ -1,9 +1,9 @@
 import type { LGraphNode } from '@comfyorg/litegraph'
 import type { INumericWidget } from '@comfyorg/litegraph/dist/types/widgets'
 
+import type { InputSpec } from '@/schemas/apiTypes'
 import type { ComfyWidgetConstructor } from '@/scripts/widgets'
 import { useSettingStore } from '@/stores/settingStore'
-import type { InputSpec } from '@/types/apiTypes'
 import { getNumberDefaults } from '@/utils/mathUtil'
 
 export const useFloatWidget = () => {
