+++ src/composables/widgets/useSeedWidget.ts
@@ -1,8 +1,8 @@
 import type { LGraphNode } from '@comfyorg/litegraph'
 
+import type { InputSpec } from '@/schemas/apiTypes'
 import type { ComfyWidgetConstructor } from '@/scripts/widgets'
 import type { ComfyApp } from '@/types'
-import type { InputSpec } from '@/types/apiTypes'
 
 import { useIntWidget } from './useIntWidget'
