+++ src/composables/widgets/useStringWidget.ts
@@ -1,9 +1,9 @@
 import type { IWidget, LGraphNode } from '@comfyorg/litegraph'
 
+import type { InputSpec } from '@/schemas/apiTypes'
 import type { ComfyWidgetConstructor } from '@/scripts/widgets'
 import { useSettingStore } from '@/stores/settingStore'
 import type { ComfyApp } from '@/types'
-import type { InputSpec } from '@/types/apiTypes'
 
 function addMultilineWidget(
   node: LGraphNode,
