+++ src/extensions/core/uploadAudio.ts
@@ -6,9 +6,9 @@ import { useNodeDragAndDrop } from '@/composables/node/useNodeDragAndDrop'
 import { useNodeFileInput } from '@/composables/node/useNodeFileInput'
 import { useNodePaste } from '@/composables/node/useNodePaste'
 import { t } from '@/i18n'
+import { ComfyNodeDef } from '@/schemas/apiTypes'
 import type { DOMWidget } from '@/scripts/domWidget'
 import { useToastStore } from '@/stores/toastStore'
-import { ComfyNodeDef } from '@/types/apiTypes'
 
 import { api } from '../../scripts/api'
 import { app } from '../../scripts/app'
