+++ src/scripts/ui/components/button.ts
@@ -1,6 +1,6 @@
 // @ts-strict-ignore
+import { Settings } from '@/schemas/apiTypes'
 import type { ComfyApp } from '@/scripts/app'
-import { Settings } from '@/types/apiTypes'
 
 import type { ComfyComponent } from '.'
 import { $el } from '../../ui'
