+++ src/scripts/ui/settings.ts
@@ -1,7 +1,7 @@
+import type { Settings } from '@/schemas/apiTypes'
 import type { ComfyApp } from '@/scripts/app'
 import { useSettingStore } from '@/stores/settingStore'
 import { useToastStore } from '@/stores/toastStore'
-import type { Settings } from '@/types/apiTypes'
 import type { SettingParams } from '@/types/settingTypes'
 
 import { ComfyDialog } from './dialog'
