+++ src/scripts/widgets.ts
@@ -14,8 +14,8 @@ import { useMarkdownWidget } from '@/composables/widgets/useMarkdownWidget'
 import { useSeedWidget } from '@/composables/widgets/useSeedWidget'
 import { useStringWidget } from '@/composables/widgets/useStringWidget'
 import { t } from '@/i18n'
+import type { InputSpec } from '@/schemas/apiTypes'
 import { useSettingStore } from '@/stores/settingStore'
-import type { InputSpec } from '@/types/apiTypes'
 
 import type { ComfyApp } from './app'
 import './domWidget'
