+++ src/services/litegraphService.ts
@@ -12,6 +12,7 @@ import { IBaseWidget, IWidget } from '@comfyorg/litegraph/dist/types/widgets'
 
 import { useNodeImage, useNodeVideo } from '@/composables/node/useNodeImage'
 import { st } from '@/i18n'
+import { ComfyNodeDef } from '@/schemas/apiTypes'
 import type { NodeId } from '@/schemas/comfyWorkflow'
 import { ANIM_PREVIEW_WIDGET, ComfyApp, app } from '@/scripts/app'
 import { $el } from '@/scripts/ui'
@@ -19,7 +20,6 @@ import { calculateImageGrid, createImageHost } from '@/scripts/ui/imagePreview'
 import { useCanvasStore } from '@/stores/graphStore'
 import { useNodeOutputStore } from '@/stores/imagePreviewStore'
 import { useToastStore } from '@/stores/toastStore'
-import { ComfyNodeDef } from '@/types/apiTypes'
 import { normalizeI18nKey } from '@/utils/formatUtil'
 import { is_all_same_aspect_ratio } from '@/utils/imageUtil'
 import { getImageTop, isImageNode, isVideoNode } from '@/utils/litegraphUtil'
