+++ src/stores/queueStore.ts
@@ -3,9 +3,6 @@ import { defineStore } from 'pinia'
 import { toRaw } from 'vue'
 import { computed, ref } from 'vue'
 
-import type { ComfyWorkflowJSON, NodeId } from '@/schemas/comfyWorkflow'
-import { api } from '@/scripts/api'
-import type { ComfyApp } from '@/scripts/app'
 import type {
   ResultItem,
   StatusWsMessageStatus,
@@ -14,7 +11,10 @@ import type {
   TaskPrompt,
   TaskStatus,
   TaskType
-} from '@/types/apiTypes'
+} from '@/schemas/apiTypes'
+import type { ComfyWorkflowJSON, NodeId } from '@/schemas/comfyWorkflow'
+import { api } from '@/scripts/api'
+import type { ComfyApp } from '@/scripts/app'
 
 // Task type used in the API.
 export type APITaskType = 'queue' | 'history'
