+++ src/stores/settingStore.ts
@@ -3,9 +3,9 @@ import { defineStore } from 'pinia'
 import type { TreeNode } from 'primevue/treenode'
 import { computed, ref } from 'vue'
 
+import type { Settings } from '@/schemas/apiTypes'
 import { api } from '@/scripts/api'
 import { app } from '@/scripts/app'
-import type { Settings } from '@/types/apiTypes'
 import type { SettingParams } from '@/types/settingTypes'
 import { buildTree } from '@/utils/treeUtil'
