+++ src/stores/userStore.ts
@@ -1,8 +1,8 @@
 import { defineStore } from 'pinia'
 import { computed, ref, watchEffect } from 'vue'
 
+import type { User as UserConfig } from '@/schemas/apiTypes'
 import { api } from '@/scripts/api'
-import type { User as UserConfig } from '@/types/apiTypes'
 
 export interface User {
   userId: string
