+++ src/utils/mathUtil.ts
@@ -1,4 +1,4 @@
-import type { InputSpec } from '@/types/apiTypes'
+import type { InputSpec } from '@/schemas/apiTypes'
 
 export function getNumberDefaults(
   inputOptions: InputSpec[1],
